Extract `fs-api` and `fs-sim` packages from consensus (test) packages.
The `fs-api` package contains the `HasFS` interface to the file system,
and it provides the `HasFS` implementation for the `IO` file system.

The `fs-sim` package contains the `HasFS` implementation for a simulated
file system, together with some test utilities. We'll have to
investigate whether the test utilities should live in a separate lib.

Some utility modules from `ouroboros-consensus` are copied into the
packages and adapated, because we can not depend on
`ouroboros-consensus`, since that would introduce circular dependencies.
In some other cases where dependencies on definitions from utility
modules were only minor, those definitions were inlined or copied from
the utility modules into the `fs-api` and `fs-sim` code.

This commit removes the core `fs-api` and `fs-sim` from
`ouroboros-consensus` and `ourobors-consensus-test`, but does not yet
fix the compilation errors that arise in the `ouroboros-consensus*`
packages. A future commit should make the `ouroboros-consensus-*`
packages depend on `fs-api` and `fs-sim`, and fix the all the
compilation errors.
