+++ fs-sim/fs-sim.cabal
@@ -1,0 +1,96 @@
+cabal-version:      3.0
+
+name:               fs-sim
+version:            0.1.0.0
+synopsis:           Simulated file systems
+description:        Simulated file systems.
+license:            Apache-2.0
+license-files:      LICENSE
+                    NOTICE
+copyright:          2019-2023 Input Output Global Inc (IOG)
+author:             IOHK Engineering Team
+maintainer:         operations@iohk.io
+category:           Testing
+build-type:         Simple
+extra-doc-files:    CHANGELOG.md
+
+source-repository head
+  type:     git
+  location: https://github.com/input-output-hk/ouroboros-network
+
+flag asserts
+  description: Enable assertions
+  manual:      False
+  default:     False
+
+library
+  hs-source-dirs:     src
+
+  exposed-modules:    System.FS.Sim.Error
+                    , System.FS.Sim.FsTree
+                    , System.FS.Sim.MockFS
+                    , System.FS.Sim.Pure
+                    , System.FS.Sim.STM
+
+  default-language:   Haskell2010
+
+  build-depends:      base        >=4.14  && <4.17
+                    , base16-bytestring
+                    , bytestring  >=0.10  && <0.12
+                    , containers  >=0.5   && <0.7
+                    , fs-api
+                    , io-classes ^>=0.3
+                    , mtl
+                    , nothunks    >=0.1.2 && <0.2
+                    , QuickCheck
+                    , strict-stm
+                    , text        >=1.2   && <1.3
+
+  ghc-options:        -Wall
+                      -Wcompat
+                      -Wincomplete-uni-patterns
+                      -Wincomplete-record-updates
+                      -Wpartial-fields
+                      -Widentities
+                      -Wredundant-constraints
+                      -Wmissing-export-lists
+                      -Wunused-packages
+
+  if flag(asserts)
+    ghc-options:      -fno-ignore-asserts
+    cpp-options:      -DENABLE_ASSERTIONS
+
+test-suite fs-sim-test
+  type:               exitcode-stdio-1.0
+  hs-source-dirs:     test
+  main-is:            Main.hs
+  other-modules:      Test.System.FS.StateMachine
+                    , Test.Util.RefEnv
+  default-language:   Haskell2010
+  build-depends:      base ^>=4.16.4.0
+                    , bifunctors
+                    , bytestring
+                    , containers
+                    , fs-api
+                    , fs-sim
+                    , generics-sop
+                    , pretty-show
+                    , QuickCheck
+                    , quickcheck-state-machine
+                    , random
+                    , tasty
+                    , tasty-quickcheck
+                    , temporary
+                    , text
+                    , tree-diff
+
+  ghc-options:        -Wall
+                      -Wcompat
+                      -Wincomplete-uni-patterns
+                      -Wincomplete-record-updates
+                      -Wpartial-fields
+                      -Widentities
+                      -Wredundant-constraints
+                      -Wmissing-export-lists
+                      -Wunused-packages
+                      -fno-ignore-asserts
